Test files
To test the Kandilli scrape part, changes to the Kandilli scrape code have been made to create mock DynamoDB tables. Using the Kandilli scrape code available on GitHub will not work for testing. There are no major issues in the text preprocessing code. The to_lower function has trouble converting "I" to "ı" in Turkish; instead, it converts it to "i".